Introduce merging function; Move utilities tests one level up so they will be executed
