+++ src/index.js
@@ -6,8 +6,7 @@ import { toInt, isObject } from './utils/unit'
 import EventsBus from './core/event/events-bus'
 
 export default class Glide {
-  /**
-   * Construct glide.
+  /**   * Construct glide.
    *
    * @param  {String} selector
    * @param  {Object} options
@@ -17,6 +16,7 @@ export default class Glide {
     this._e = new EventsBus()
 
     this.disabled = false
+
     this.selector = selector
     this.settings = Object.assign({}, defaults, options)
     this.index = this.settings.startAt
