+++ tests/unit/string.test.js
@@ -1,4 +1,4 @@
-import { ucfirst } from '../../../src/utils/string'
+import { ucfirst } from '../../src/utils/string'
 
 describe('Function', () => {
   test('`ucfirst` should capitalize passed string', () => {
